requirements.txt
gradio==4.15.0
requests==2.31.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
mediapipe==0.10.9
pillow==10.2.0
toml==0.10.
grad-cam==1.5.0
soundfile==0.12.1
speechbrain==0.5.16
jieba==0.42.1
onnxruntime==1.16.3
kaldi-native-fbank==1.18.6
transformers==4.37.1
moviepy
openai